# perf fix for Ipv4Instance creation