*Description*:
Speed up the creation of an `Ipv4Instance` from a `sockaddr_in`.
I noticed in a CPU profile that this constructor was taking a surprisingly large amount of time.
This change includes a new performance benchmark test.

Before:
```
Benchmark                   Time           CPU Iterations
----------------------------------------------------------
Ipv4InstanceCreate        474 ns        474 ns    1470706
```
After:
```
----------------------------------------------------------
Benchmark                   Time           CPU Iterations
----------------------------------------------------------
Ipv4InstanceCreate         58 ns         58 ns   12214700
```

*Risk Level*: medium

*Testing*: unit tests plus new performance benchmark test

*Docs Changes*: N/A
*Release Notes*: N/A

Signed-off-by: Brian Pane <brianp+github@brianp.net>